Final db flush only in dev mode for now, will need separate dev module

+++ modules/Core/Controller/App.php
@@ -189,8 +189,11 @@ class App {
     $response->send();
 
     $kernel->terminate($this->getRequest(), $response);
-    // Execute any queries left behind
-    Db::flushAll();
+    
+    if($this->inDevelopment()) {
+      // Execute any queries left behind
+      Db::flushAll();
+    }
 
     return $this;
   }
